Ignore stderr when we call JavaFx provider first
The stderr maybe benign, but in its presence our process fails.
We should first try parse the stdout and if we could parse out a
correct authorization code from the stdout, then treat the error
more seriously.  Otherwise even if there are stderr, but the code
is retrieved, we should simply ignore it.