Support for Java 9+ version strings
Java 9 and later uses a new format for version strings. This change
uses the new format first and if no match is found falls back to the
previous version string format.

Here is a reference to the JDK work item for the new format: https://bugs.openjdk.java.net/browse/JDK-8061493